chatbot/requirements.txt
flask==2.3.3
flask-cors==3.0.10
Werkzeug==2.3.7

pillow==9.0.0
numpy==1.22.0